% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.r
\name{dNeedle}
\alias{dNeedle}
\title{Draw Needles}
\usage{
dNeedle(sf, study = " ")
}
\arguments{
\item{sf}{output of \code{sampleFrac}}

\item{study}{character string specifying study ID}
}
\description{
Create an html base64 string from a png graphic to draw needles for current sample sizes.  Uses colors set by call to \code{sethreportOptions}.
}
