% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.r
\name{gethreportOption}
\alias{gethreportOption}
\title{Get hreport Options}
\usage{
gethreportOption(opts = NULL, study = " ")
}
\arguments{
\item{opts}{character vector containing list of option names to retrieve.  If only one element, the result is a scalar, otherwise a list.  If \code{opts} is not specified, a list with all current option settings is returned.}

\item{study}{character string specifying an optional study designation}
}
\description{
Get hreport options, assigning default values of unspecified options.
}
