% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hreport-package.r
\docType{package}
\name{hreport-package}
\alias{hreport}
\alias{hreport-package}
\title{Interactive Graphical HTML Reports for Clinical Trials}
\description{
Contains many functions useful for
    monitoring and reporting the results of clinical trials and other
    experiments in which treatments are compared.  HTML is
    used to typeset the resulting reports, recommended to be in the
    context of 'knitr' HTML notebooks. The 'Hmisc', 'plotly', 'ggplot2', and
    'lattice' packages are used by 'greport' for high-level graphics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hbiostat.org/R/hreport}
  \item \url{https://github.com/harrelfe/hreport}
}

}
\author{
Frank E Harrell Jr \email{f.harrell@vanderbilt.edu}
}
