% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nriskReport.r
\name{nriskReport}
\alias{nriskReport}
\title{Number at Risk Report}
\usage{
nriskReport(
  formula,
  time0 = "randomization",
  data = NULL,
  subset = NULL,
  na.action = na.retain,
  study = " ",
  ylab = "Number Followed",
  head = NULL,
  tail = NULL,
  h = 400,
  w = 700
)
}
\arguments{
\item{formula}{a formula with time on the left hand side, and with variables on the right side being possible stratification variables.  If no stratification put \code{1} as the right hand side.  Specify unique participant IDs by including a term \code{id()} if participantss have more than one observation.}

\item{time0}{a character string defining the meaning of time zero in follow-up.  Default is \code{"randomization"}.}

\item{data}{data frame}

\item{subset}{a subsetting epression for the entire analysis}

\item{na.action}{a NA handling function for data frames, default is \code{na.retain}}

\item{study}{character string identifying the study; used in multi-study reports or where distinct patient strata are analyzed separately.  Used to fetch the study-specific metadata stored by \code{\link{sethreportOption}}.  Single study reports just use \code{study=' '}.}

\item{ylab}{character string if you want to override \code{"Number Followed"}}

\item{head}{character string.  Specifies initial text in the figure caption, otherwise a default is used}

\item{tail}{optional character string.  Specifies final text in the figure caption, e.g., what might have been put in a footnote in an ordinary text page.  This appears just before any needles.}

\item{h}{numeric.  Height of plot, in pixels}

\item{w}{numeric.  Width of plot}
}
\description{
Graph number of participants at risk
}
\details{
\code{nriskReport} generates several plots where stratification is by the cross-classifications of all right-hand-side variables other than the time variable and an \code{id} variable.  The first plot depicts the number at risk as a function of follow-up time.  It is assumed that this function is only run on randomized participants.  If an \code{id} variable is present but stratification variables are not, other plots are also produced: a histogram of the number of contacts per participant, a histogram of times at which participants have contacts, the average number of contacts as a function of elapsed time, and a histogram showing the distribution of the longest gap between contacts over participants.
}
\examples{
# See test.Rnw in tests directory
}
