% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.r
\name{pdumpit}
\alias{pdumpit}
\title{Print to File for Debugging}
\usage{
pdumpit(x, txt = as.character(substitute(x)))
}
\arguments{
\item{x}{input to \code{prn}}

\item{txt}{text label, defaults to name of \code{x} argument}
}
\description{
If \code{options(dumpfile="...")} is set, uses \code{Hmisc::prn()} to print objects for debugging
}
\author{
Frank Harrell
}
