% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.r
\name{sethreportOption}
\alias{sethreportOption}
\title{Set hreport Options}
\usage{
sethreportOption(..., study = " ")
}
\arguments{
\item{\dots}{a series of options for which non-default values are desired:
\itemize{
 \item{\code{tx.pch}:}{symbols corresponding to treatments}
 \item{\code{tx.col}:}{colors corresponding to treatments}
 \item{\code{tx.linecol}:}{colors for lines in line plots}
 \item{\code{nontx.col}:}{colors for categories other than treatments}
 \item{\code{tx.lty}:}{line types corresponding to treatments}
 \item{\code{tx.lwd}:}{line widths corresponding to treatments}
 \item{\code{tx.var}:}{character string name of treatment variable}
 \item{\code{er.col}:}{2-vector with names \code{"enrolled","randomized"} containing colors to use for enrolled and randomized in needle displays}
 \item{\code{alpha.f}:}{single numeric specifying alpha adjustment to be applied to all colors.  Default is 1 (no adjustment)}
 \item{\code{denom}:}{named vector with overall sample sizes}
}}
}
\description{
Set hreport Options
}
