% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survReport.r
\name{survReport}
\alias{survReport}
\title{Survival Report}
\usage{
survReport(
  formula,
  data = NULL,
  subset = NULL,
  na.action = na.retain,
  study = " ",
  ylab = NULL,
  what = c("1-S", "S"),
  conf = c("bands", "none"),
  cause = NULL,
  head = NULL,
  tail = NULL,
  h = 425,
  w = 675,
  mylim = NULL,
  aehaz = TRUE,
  times = NULL,
  opts = NULL,
  ...
)
}
\arguments{
\item{formula}{a formula with survival (\code{Surv}) objects on the left hand side and an optional stratification factor on the right (or \code{1} if none).  The survival object component variables should be labeled; these labels are used for graph annotation.  If any of the \code{Surv} objects are competing risk objects (see \code{\link[survival]{Surv}}), event labels come from the factor levels in the variable that was the second argument to \code{Surv}, and the first factor level must correspond to right-censored observations.}

\item{data}{data.frame}

\item{subset}{optional subsetting criteria}

\item{na.action}{function for handling \code{NA}s while creating a data frame}

\item{study}{character string identifying the study; used in multi-study reports or where distinct patient strata are analyzed separately.  Used to fetch the study-specific metadata stored by \code{\link{sethreportOption}}.  Single study reports just use \code{study=' '}.}

\item{ylab}{character. Passed to \code{\link[rms]{survplotp.npsurv}} as the \code{ylab} argument.  Constructed by default.}

\item{what}{\code{"1-S"} (the default) to plot cumulative incidence functions or \code{"S"} to plot cumulative survival functions.  If any of the survival time objects on the left hand side are competing risk objects, \code{"S"} may not be used.}

\item{conf}{character. See \code{\link[rms]{survplotp.npsurv}}.}

\item{cause}{character vector or list.  If a vector, every \code{Surv} term on the left hand side of \code{formula} will have cumulative incidence plotted for all causes that appear in \code{cause}.  If a list, the list elements must correspond to the \code{Surv} terms in order, and specify which causes to display from the corresponding \code{Surv} object.  When \code{cause} is a list and one of its elements contains more than one character string, or when \code{cause} is a vector and for one \code{Surv} object it matches multiple causes, \code{survReport} produces more plots than there are \code{Surv} objects.}

\item{head}{character string.  Specifies initial text in the figure caption, otherwise a default is used.}

\item{tail}{optional character string.  Specifies final text in the figure caption, e.g., what might have been put in a footnote in an ordinary text page.  This appears just before any needles.}

\item{h}{numeric. Height of plots in pixels.}

\item{w}{numeric. Width of plots in pixels.}

\item{mylim}{numeric 2-vector.  Used to force expansion of computed y-axis limits.  See \code{survplotp}.}

\item{aehaz}{logical.  Set to \code{FALSE} to not print number of events and hazard rate on plots.}

\item{times}{numeric vector.  If specified, prints cumulative incidence probabilities at those times on the plots.}

\item{opts}{list.  A list specifying arguments to \code{survReport} that override any other arguments.  This is useful when making a long series of \code{survReport} calls with the same options, as the options can be defined up front in a list.}

\item{\dots}{ignored}
}
\description{
Generate a Survival Report with Kaplan-Meier Estimates
}
\examples{
# See tests directory test.Rnw for a live example
\dontrun{
  set.seed(1)
  n <- 400
  dat <- data.frame(t1=runif(n, 2, 5), t2=runif(n, 2, 5),
                    e1=rbinom(n, 1, .5), e2=rbinom(n, 1, .5),
                    treat=sample(c('a','b'), n, TRUE))
  dat <- upData(dat,
                labels=c(t1='Time to operation',
                         t2='Time to rehospitalization',
                         e1='Operation', e2='Hospitalization',
                         treat='Treatment')
                units=c(t1='year', t2='year'))
  survReport(Surv(t1, e1) + Surv(t2, e2) ~ treat, data=dat)

  dat <- upData(dat, labels=c(t1='Follow-up Time', t2='Time'),
                cause=factor(sample(c('death','MI','censor'), n, TRUE),
                             c('censor', 'MI', 'death')))
  survReport(Surv(t1, cause) ~ treat, cause='death', data=dat)
  survReport(Surv(t1, cause) + Surv(t2, cause) ~ treat,
             cause=list(c('death', 'MI'), 'death'), data=dat)
  # Two plots for t1, one plot for t2
}
}
\references{
Boers M (2004): Null bar and null zone are better than the error bar to compare group means in graphs.  J Clin Epi 57:712-715.
}
